requirements.txt
Flask
rembg
Pillow